examples: zephyr: lightdb/observe: increase test timeout to 5s
Test timeout was set to 2s only, between setting lightdb /counter value and
expecting console logs to reflect that change. Console logs have a delay of
up to 1s (configurable in Kconfig), which leaves only 1s for the network
roundtrip, which is very fragile to not meet the timeout requirements in
CI.

Increase 2s timeout to 5s, so that there are 4s for the network roundtrip
and reaching target. This is 4x longer, so it should be long enough to
always meet timing requirements.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>