settings: GET settings after each registration
Previously, when a user registered a setting with the SDK, the value
of that setting would not be updated until the next time it was
changed in the console. Instead, when a user registers a setting
with the settings service, we should request the settings object
from the backend so that we can update the callbck with the latest
setting value.

Signed-off-by: Sam Friedman <sam@golioth.io>