coap_client: avoid log loop of doom
Previously, if the logging thread failed to send a message because
the request queue was full, that would cause the coap_client to
log a warning. The logging thread would immediately attempt to
enqueue a message to log the warning to the cloud, but that would
fail, generating another warning, which the logging thread would
again attempt enqueue. This causes an infinite loop of warnings
and prevents the request queue from ever draining.

Removing the warning when the request queue is full prevents us
from entering this doom loop.

Signed-off-by: Sam Friedman <sam@golioth.io>