zephyr: fw_update: replace pytest delay
With the arbitrary delay of 2 seconds, the twister test for fw_update was
frequently failing for the nRF52840 as the device flash wasn't yet ready to
store credentials. Waiting for the app start log message resolves the
timing issue.

Sometimes the app start log message is missed by the mimxrt1024_evk. Pytest
will wait 10 seconds for the message and continue if not found to work
around this issue.

Signed-off-by: Mike Szczys <mike@golioth.io>