pytest-hil: respect timeout when waiting for serial input
The pytest-hil package uses serial.readline() in a loop to wait for
strings. Each iteration through the loop, we check for a timeout.
But serial.readline() itself has no timeout and so if we receive
no input at all, we will never check the timeout and hang forever.
This happens sometimes when the serial port gets locked up (e.g.
on the mimxrt1024).

Signed-off-by: Sam Friedman <sam@golioth.io>