zephyr: drop POSIX_API dependency
Don't select POSIX_API, which was a hard requirements for libcoap. Use
zsock_* APIs, so that NET_SOCKETS_POSIX_NAMES is optional as well.

Change and reorder some of the header include list to solve include order
conflicts.

Set default value for POSIX_MAX_FDS to 16, which was exactly that with
POSIX_API being enabled.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>