port: esp_idf: convert golioth log level to esp_idf log_level
We had been passing `enum golioth_debug_log_level` to a function
that takes arguments of type `esp_log_level_t`. These values happen
to be the same, but there is no guarantee of that, and either Golioth
or Espressif could change the enum at any point and break compatibility.
Now, we explicitly convert between the enums per value.

Signed-off-by: Sam Friedman <sam@golioth.io>