Use @retval instead of @return when relevant
`@retval` expects the first token to be a literal value, and renders it
as code in the docs. This looks a little nicer than `@return` both in
our generated docs and most code completion tools.

This PR changes the return value description to `@retval` for all
functions that only list return values, but keeps the ones that still
have a mix as-is.

Signed-off-by: Trond Snekvik <trond@golioth.io>