stream: remove set functions for individual, non-object types
The functions for setting individual values in lightdb stream rely
on behavior that will be complex to support by default in conjunction
with new development around data ingestion.

Additionally, these functions are not believed to be useful beyond
trivial examples. In the vast majority of real use-cases, it is
expected that applications will want to send multiple values in one
message, even if just timestamp + sensor reading. This requires using
the object APIs.

Signed-off-by: Sam Friedman <sam@golioth.io>