tests: hil: pytest-hil: increase delay after programming
Now that we erase the boards between test runs, we need to delay
at the start of each test to give time for the settings partition
to be reformatted before we attempt to store settings to flash.

This is a temporary fix until a more robust solution is in place.

Signed-off-by: Sam Friedman <sam@golioth.io>