workflows: esp-idf: erase boards after tests
If a board had stored credentials when a test runs, the new credentials
may not be used immediatel in some cases. For instance, if a WiFi-based
board has already established a connection when new credentials are set,
they will not be immediately used since a connetion is currently in place.

This commit erases board flash at the end of test runs to ensure all
applications will be provisioned with the expected AP/Golioth device
credentials the next time a test runs.

Signed-off-by: Mike Szczys <mike@golioth.io>