examples: zephyr: extend twister test timeouts
The nrf9160 was sometimes hitting twister timeouts before it was
able to establish a cell connection. This overall timeout was
shorter than the sum of the individual timeouts in the pytest file.
This timeout is a last case timeout to detect a hung board, not to
ensure compliance with any deadline, so it doesn't need to be
particularly tight. Extending it to 180 seconds should give plenty
of time for the board to connect to the cell network and complete
the test.

Signed-off-by: Sam Friedman <sam@golioth.io>