zephyr: backend logging: enable/disable using a context flag
The Zephyr port was using log_backend_enable()/disable() to prevent logs
from generating CoAP events when not connected to Golioth. This interfere
with applications using the same functions to enable or disable backend
logging at runtime.

This commit adds an `enabled` flag to the backend context which is used to
determine if a log should be processed. This way, the SDK can
disable/enable backend logging when the connection status changes, while
retaining the ability for applications to control the backend through
normal Zephyr logging management functions.

Signed-off-by: Mike Szczys <mike@golioth.io>