zephyr: certificate_provisioning: pytest: trigger on different message
The first few seconds of log messages are missed by the mimxrt1024rt board
due the autostart of the Golioth logging backend (any backend that
autostarts causes the same behavior on this board). Instead of looking for
the first log message, trigger of the `<inf>` message, which indicates that
the app has started running.

Signed-off-by: Mike Szczys <mike@golioth.io>