examples: esp-idf: rpc: wait for RPC registered before calling in test
Some failures in CI were caused by the test script attempting to call
the RPC before the device had established an observation on the RPC
endpoint. This commit adds a new message that the device prints when
it has established the observation, and updates the test script to
wait for that message.

Signed-off-by: Sam Friedman <sam@golioth.io>