zephyr_coap_req: directly index into header for coap response code
The Zephyr CoAP stack filters out unknown CoAP codes, but due to a
bug, it incorrectly filters out the TOO_MANY_REQUESTS message.
We log an error when we receive an unknown code, which results
in us getting into a loop where the server responds with
TOO_MANY_REQUESTS, the device logs an error in response, and
the loop repeats.

This bug is fixed in Zephyr upstream, but not in the latest
release of NCS. To workaraound the issue in NCS, we can
retrieve the code by directly indexing into the header,
bypassing the filtering done by the Zephyr CoAP stack.
When NCS is updated to include the fix, we will revert this
change.

Signed-off-by: Sam Friedman <sam@golioth.io>