zephyr: delete TLS PSK credentials before setting them
Delete the PSK and PSK-ID from the TLS credentials management before trying
to set them.

In the case that runtime credentials are set after boot, but before the
Golioth Client connects, an -EEXIST error (-17; file exists) will be thrown
once the Client connects because the tags have already been set in TLS
credentials management.

This is commonly encountered with automated testing. However, this will
also ensure that the credentials passed by golioth_client_create() are
what is actually used.

Signed-off-by: Mike Szczys <mike@golioth.io>