examples: zephyr: Fix get_temperature with temp0 DT alias
When a temp0 devicetree alias is added to the board overlay, the
get_temperature function was looping through all sensor channels, always
returning the result of reading the last one, even if it was not
a supported channel.

This fixes the function so that it returns after the first successful
read from a temperature channel.

Signed-off-by: Chris Wilson <chris@cgnd.dev>