zephyr: coap: skip free for observations
Observations are stored in the golioth client struct so that the data may
be used to reestablish observations upon reconnect. When the Golioth client
is stopped, all requests in the client->coap_reqs linked list are freed.
This causes the client address in the stored observation to change, leading
to a crash when trying to reestablish. This commit skips freeing
observations so the data remains intact. This memory should only be freed
when cancelling an observation.

Signed-off-by: Mike Szczys <mike@golioth.io>