zephyr: coap: use client slots for observation messages
Use the Golioth client observation slots as the request message. This
solves an issue when first establishing observations where the address of
the request message is not a slot in the client, however, reestablished
observations do use these slots. We need to track the pointer to the
request messages as a handle for removing them from the coap_reqs list when
cancelling requests.

Signed-off-by: Mike Szczys <mike@golioth.io>