libcoap: cancel all observations
Observation updates are received from the server as unsolicited requests.
The token will always match the token used when the observation was
registered. The client observations list stores a copy of the original
request message that contains the token to compare against, as well as the
callback to notify.

Observations are cancelled by settings the client observations slot as
!in_use, then sending an "eager release" request to the server using the
original token, path, etc.

Signed-off-by: Mike Szczys <mike@golioth.io>