ci: cancel previous in progressi HIL runs
We have a limited number of machines that can run our HIL tests,
and during periods of increased activity a long backlog can form
of tests waiting for runners to become available. To mitigate this,
we have been manually cancelling older runs of HIL tests that
are no longer relevant (because a newer reivision of the PR
exists), in order to free up more runners. This PR updates our CI
HIL test workflow so that cancelling older runs of the workflow
on outdated revisions now occurs automatically.

Signed-off-by: Sam Friedman <sam@golioth.io>