tests: hil: rpc: wait for RPC establishment before sending RPCs
In the observation_repeat_restart test, we were waiting for the device
to establish a connection to Golioth before the test script sends an
RPC to the device. This creates a small window between when the device
establishes a connection and when it establishes an observation on the
RPC path where attempts to send an RPC will fail. Instead, we should
wait for the device to report that is successfully established an
observation before the test script attempts to send any RPCs to the
device. This matches the behavior in other parts of the test script.

Signed-off-by: Sam Friedman <sam@golioth.io>