zephyr: coap: prevent resend backend logs
When a network connection is unstable, coap packets may be resent. Logging
this information is useful, but if backend logging is enabled, these get
added to the network queue, further compounding the problem. Check to
ensure backend logging is disabled before creating a log message about a
coap request being resent.

Signed-off-by: Mike Szczys <mike@golioth.io>