tests: hil: ota: wait for successful status report
Previously, the test script was waiting on a string printed by the
device _before_ it reports it status, then waiting for a specific delay
before retrieiving the reported status from the backend. Instead, the
test script now waits for a string that is printed by the device after
it receives an acknowledgement from the server that the status update
has been posted successfully. This makes the timing of the test more
explicit and should reduce timing related failures in CI.

Signed-off-by: Sam Friedman <sam@golioth.io>