ci: run twister samples with 'native_sim'
Apply patches for NSOS support. Create new test job (instead of reusing
template) as there are many steps that are different for native_sim. It
also makes sense to build and run in single invocation, instead of keeping
build and test separate.

Clone 'picolibc' module in CI, as that is required for all toolchains other
than Zephyr SDK (in case of native_sim it is host toolchain with picolibc
as C standard library).

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>