coap_blockwise: use non-blocking APIs for CoAP operations
The blocking APIs in the CoAP client create and destroy a semaphore each
time they are called. For a complete blockwise transfer, we could call
these APIs thousands of times. To avoid creating and destroying a
semaphore each time, we instead manage the blocking in the blockwise
transfers module, which now uses the non-blocking APIs. This allows
us to create and destroy the semaphore just once for the complete
blockwise transfer, rather than for each block.

Also, rename the `status` member of the context struct to
`response_status` to better reflect how it is used.

Signed-off-by: Sam Friedman <sam@golioth.io>