blockwise: use local structs for client callback args
The `blockwise_transfer` context struct is being used both to store the
parameters for the whole transfer as well as for storing variables that are
used for communicating between callbacks from the coap client and the
blockwise transfer module. This leads to some confusion over the purpose of
the members of the context struct, and causes us to hold additional
allocated memory.

Instead, the coap client callbacks (from `golioth_coap_client_get_block()`
and `golioth_coap_client_set_block()` use stack local structs to carry
information. This helps both readability and memory usage by limiting
the lifetime and scope of those parameters to only the duration in
which they are used.

Note that this commit switches to using the stack local structs, but there
are still temporary variables held in the overall context struct. These
will be removed and reworked in a subsequent commit to keep changes small.

Signed-off-by: Sam Friedman <sam@golioth.io>