coap_blockwise: pass block buffer size to read_block callback
The block size should be negotiated between client and server, so
the user application cannot rely on any constant value to know how
large the block buffer is. While we aren't actually negotiating the
block size yet, pass the constant buffer size to the user callback
for reading blocks so that callbacks can be written to be ready for
negotiation in the future.

Signed-off-by: Sam Friedman <sam@golioth.io>