coap_blockwise: track individual block size on stack
Previously, we were storing the size of any individual block in the
blockwise transfer context struct. This information is only needed
for the duration of downloading a single block and calling the write
block callback (or calling the read block callback and uploading a
single block). So it can be kept local on the stack.

Also, we were not storing the negotiated block size anywhere, instead
relying on the hardcoded value set through kconfig. While we are still
not actually negotiating the block size, we now store the negotiated
size (here hardcoded) in then context struct, and use it for bounds
checking. This will set us up well for the future when we implement
block size negotiation.

Signed-off-by: Sam Friedman <sam@golioth.io>