hil: ota: wait for firmware update state to propagate
Introduces a short wait (2 seconds) for firmware update state changes to
propagate. State changes are not guaranteed to be immediately reflected for
a device after an OK response is received.

In the future, this stop-gap solution should be replaced by either polling
in this test or synchronous behavior on the server. Currently the on-device
test is waiting 7s between state updates, so the wait in this test must
remain less than to avoid missing a state update.

Signed-off-by: Daniel Mangum <georgedanielmangum@gmail.com>