workflows: zephyr: patch before build
Patches that add blackmagicprobe runner support for the RAK5010 are
necessary on the build job or they will not be available on the test job.
This patch will not be necessary once Zephyr v3.7.0 is available, but we
should still patch on the build step to avoid issues in the future.

Signed-off-by: Mike Szczys <mike@golioth.io>