coap: zephyr: stop client before destroy
If the client is running when golioth_client_destroy() is called it needs
to be stopped before freeing the client.

When the client is stopped, the logging backend will be disabled,
preventing log messages from trying to access the client after it is freed.
The linked list of requests (client->coap_reqs) will then also be freed as
part of the stop process, ensuring there isn't any allocated OBSERVE or
POST/POST_BLOCK memory leaked.

Signed-off-by: Mike Szczys <mike@golioth.io>