coap: libcoap: stop client before destroy
If the client is running when golioth_client_destroy() is called it needs
to be stopped before freeing the client.

When the client is stopped, golioth_sys_client_disconnected() is now called
which in turn sets golioth_debug_set_cloud_log_enabled() to false. This
prevents attempts to send log messages to Golioth when the client is not
connected.

Signed-off-by: Mike Szczys <mike@golioth.io>