examples: zephyr: cert_provisioning: use file settings backend
By default, we use the NVS backend to persist settings for things like
PSK and WiFi credentials. In the cert provisioning sample, we have a
file system enabled in order to store credentials, which conflicts with
the settings NVS partition. To avoid the conflict, we can use a file as
the backend for persistent settings storage.

Signed-off-by: Sam Friedman <sam@golioth.io>