examples: zephyr: cert_provisioning: Enable twister test for wifi boards
Enable Twister testing for esp32_devkitc_wrover and nRF52840dk. These
boards couldn't be tested before because they required the settings
subsystem to be enabled to get WiFi credentials. Now that this sample
uses the fileystem settings storage backend, we can store WiFi
credentials alongside the certificates in the filesystem.

Signed-off-by: Sam Friedman <sam@golioth.io>