ci: use esptool.py from pip instead of hal_espressif
By default, the esp32 west runner looks in the espressif HAL for
the esptool.py script. Because we don't want to checkout the HAL
during twister runs in CI (to save time), instead specify the path
for esptool.py so that the runner will use the version installed
from pip in the Docker container.

Signed-off-by: Sam Friedman <sam@golioth.io>