coap: zephyr: add log level zephyr coap client
The coap_client_zephyr.c file registers for logging but it does not pass a
log level. This meant that enabling debug logging for the Golioth Firmware
SDK did not include enabling logs from the zephyr client. This commit adds
GOLIOTH_DEBUG_DEFAULT_LOG_LEVEL as the default log level.

Signed-off-by: Mike Szczys <mike@golioth.io>