ci: linux: switch from lcov to gcovr
Zephyr tests and samples coverage is handled using gcovr. Switch to that
tool for Linux tests as well for improved consistency. This also allows to
combine coverage results for Linux and Zephyr.

Use gcovr tool to generate JSON output for each test and put most of the
coverage report generation inside 'coverage' job:
 * HTML report (instead of lcov -> genhtml in case of Linux)
 * single PR comment from XML output (instead of having two comments, each
   with slightly different format)
 * GH Actions step summary that includes Linux platform coverage (which was
   not generated before for Linux)

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>