coap: block download: add block_buffer_len, rename block_size
Add a `block_buffer_len` param to the block download callback. This
indicates the length of the data present in the `block_buffer`, which will
most often be a number smaller than the `block_size` on the final block.

Rename `block_size` to `negotiated_block_size`. This param passes the
size of the coap block. This value indicates the largest possible length of
data in the block_buffer. Use `block_size * block_idx` to calculate a
memory offset.

Now that the block download callback includes both `negotiated_block_size`
and `block_buffer_len`, we can directly use `negotiated_block_size` for
calculating offset, and `block_buffer_len` for tracking the number of bytes
received.

Signed-off-by: Mike Szczys <mike@golioth.io>