zephyr: logging: fix off-by-one in log message length limit
Fix an off-by-one error when writing a null terminator to the end of a log
message buffer before sending to Golioth.

When a log message is longer than the setting for the  maximum Golioth
backend log buffer (CONFIG_LOG_BACKEND_GOLIOTH_MAX_LOG_STRING_SIZE), the
message is truncated by writing a null terminator onto the end of the
buffer. The index for this null terminator was using a counter that had
already been incremented past the end of the buffer array. Instead, test
for an index pointing to the final array position and do not
write/increment.

Signed-off-by: Mike Szczys <mike@golioth.io>