ci: zephyr: execute (real hardware) HIL tests in matrix jobs
So far bash 'for' loop was used to iterate through all tests, then build
them and run pytest. This produced massive amount of logs and as a result
was very hard to find a problem when such occurred.

Add job (matrix) that dynamically generates matrix based on the list of
tests to execute. Then use matrix strategy to build those tests and run
pytest on HIL.

Split building and runtime execution into separate steps, so that it
improves:
 * visual feedback in GitHub Actions UI when investigating failed jobs
 * no $EXITCODE handling
 * faster execution of "Re-run failed jobs" when only single test on a
   single platform has failed (it makes it more likely to pass)

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>