coap: libcoap: use negotiated block size when adding block1
Use the negotiated block size when initializing the block1 option of the
coap packet.

This moves the use of the compiled-in
CONFIG_GOLIOTH_BLOCKWISE_UPLOAD_MAX_BLOCK_SIZE value to the initialization
step of the negotiated value, and allows for the negotiated value to be
updated based on the server preference, then used for all subsequent coap
requests in the block upload process.

Signed-off-by: Mike Szczys <mike@golioth.io>