coap: zephyr: call callbacks when cancelling requests
The stored requests list is dropped when the client stops and starts. Any
callbacks associated with these requests are never called.

Instead, call the callback (if any) for each request being cancelled, using
the cancellation reason already being passed as the status code given to
the callback. This way, any callback checking for status == GOLIOTH_OK will
already be able to handle the error case.

This resolves a possible deadlock if the client stops and starts during
block download.

Signed-off-by: Mike Szczys <mike@golioth.io>