coap: zephyr: break after each case so keepalive is serviced
When a response is received, the keepalive timer should always be serviced.
This PR removes the parts of program flow that skipped over that step.

On the response callback, observations were returning after the callback
instead of using break. This intentionally prevented observations from
being freed, which is now happening just before the free command at the
bottom of the function.

Signed-off-by: Mike Szczys <mike@golioth.io>