hil: zephyr: increase log buffer to prevent dropped logs
When running the settings HIL test for Zephyr, logs are often dropped
because of the number of settings being reported on first connect. This
often results in the "Connected" message being dropped which causes the
test to fail.

Double the size of the Zephyr log buffer to 2048 to prevent logs from being
dropped.

Signed-off-by: Mike Szczys <mike@golioth.io>