coap: zephyr: do not call callbacks for cancelled observations
The two ways in which observations are cancelled are when a user explicitly
cancels them, or the Golioth client stops. Since the user knows they are
being cancelled, and the client will automatically re-establish
observations when it starts again, there is no reason to call an
observation callback during the cancellation step.

Signed-off-by: Mike Szczys <mike@golioth.io>