ci: run twister tests using with NCS when using nrf9160dk
So far twister build stage was run either with Zephyr or NCS fork of
Zephyr. Twister test stage was run always with Zephyr.

This approach is not compatible anymore (since there are obviously two
different twister implementations) after switching to Zephyr 4.0.0 and NCS
2.8.0.

Introduce '.ci-west-ncs-twister.yml' and allow 'nrf' and 'zephyr'
repositories to make tests work with twister. Use it to initialize SDK with
NCS and NCS' fork of Zephyr as dependencies, so that the same twister
implementation is used for build and testing.

Fixes: 7aa069cdc22f ("bump NCS to 2.8.0")
Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>