examples: zephyr: use 32 as default PSK max length
Using a PSK that is longer than 32 bytes does not work because the
default MBEDTLS_PSK_MAX_LEN is set to 32 bytes. In the event that
MBEDTLS_BUILTIN is set, the default will be 32. However, if
MBEDTLS_BUILTIN is not set, as is the case when NRF_SECURITY is set, the
default will be 64, which is larger than the default of 32 in mbedtls
(when not using TLS 1.3). If a PSK with length greater than 32 is used,
we will not alert that the PSK is too large, but the connection will
fail due to the mbedtls max length.

This updates to make the default match the 32 bytes used by mbedtls.
This ensures that users will not encounter connection issues unless they
have modified the max length, which now has a disclaimer that updating
MBEDTLS_PSK_MAX_LEN may be required as well.

Signed-off-by: Daniel Mangum <georgedanielmangum@gmail.com>