fw_update: add resume download for firmware update module
When an error occurs during a firmware block download, the download will
now resume from that block instead of starting from the beginning.

Each block will be included in a sha256 hash calculation during the
download and compared to the server-supplied hash to verify a successful
download. This comparison is made while the data is in RAM, before it is
written to flash.

If a flash write error occurs during block download, the firmware download
is considered a failure and will not try to resume from that block.

Signed-off-by: Mike Szczys <mike@golioth.io>