fw_update: store a copy of the target component from manifest
When the manifest is received, the target component is found and a pointer
is stored to use during the fw_update process. This means we cannot receive
another manifest file while the download is running or the pointer will be
changed by the `on_ota_manifest` callback while the thread is still using
it for downloading.

This PR stores a copy of the target component to ensure it cannot be
changed during download. The stored config and the stored component
are grouped together into a `fw_update_component_ctx` struct.

Future work will build on this to allow a new manifest to be received
without disrupting the download process. We will be able to stop dropping
received manifests (currently based on reported state) and have more
flexibility in what states we report. Eventually this will allow an update
to be cancelled while in progress.

Signed-off-by: Mie Szczys <mike@golioth.io>