fw_update: add retries when reporting state/reason
Add multiple retries with backoff when reporting firmware update
state/reason to the server.

We most commonly see the firmware report report fail when the mbox is full
after a device reboot. Add in a backoff retry will give time for the cached
requests to process before retrying to send the report.

Signed-off-by: Mike Szczys <mike@golioth.io>