fw_update: separate report from observe
Separate the reporting from the observation registration when the fw_update
thread starts running. The IDLE/READY state will now be reported at the top
of every loop.

Instead of reporting IDLE twice in a row after a successful update, report
success as the end of the UPDATING state. IDLE will then be reported at the
top of the loop and both will be shown in the OTA history for the device.

Signed-off-by: Mike Szczys <mike@golioth.io>