rpc: encode query method parameters as a list instead of map
When encoding information for the "query" method, put parameter details as
a list of values instead of a map so that order is retained.

Known issue: param name and type will appear as separate line items of the
list:

```yaml
{
  "multiply": [
    "int_a",
    2,
    "int_b",
    2
  ],
  "repeat_word": [
    "word",
    3,
    "multiple",
    2
  ]
}
```

Although it is possible to encode maps as list items, the server is
rejecting such encoding as invalid. This needs further investigation.

Signed-off-by: Mike Szczys <mike@golioth.io>