fw_update: add variables and functions to track retries with backoff
Record the timestamp at which a download fails and the desired backoff
duration in the component context. Add functions to set these variables
after each failure, check them to see if the backoff duration has expired,
and to reset them at the beginning and end of a component download
operation.

Signed-off-by: Mike Szczys <mike@golioth.io>