fw_download: implement retry with backoff for firmware update
When awaiting a retry, use the timeout value of the manifest semaphore to
wait until the appropriate backoff duration has passed before retrying a
download. This allows a new manifest to be received and acted upon while
awaiting a retry.

Set up the component for a retry as soon as the download step has finished
(whether it is a success or failure). The retry continues to be in effect
until all validation and other steps of the fw_update process have been
confirmed. Finally, reset the backoff values to indicate a retry
is not necessary.

Signed-off-by: Mike Szczys <mike@golioth.io>