fw_update: add backoff awareness to received_new_target_component()
Move the mutex lock/unlock functions inside of
received_new_target_component() as it should never be called without
guarding the manifest pointer which is also updated on another thread by
the on_ota_manifest callback.

Check stored target component against what was received with manifest and
use along with the backoff_duration_ms time value in the context to detect
when a component is already being downloaded.

Reset the backoff values in the context when a new component is received
from the manifest.

Signed-off-by: Mike Szczys <mike@golioth.io>