ports: move golioth_sys_sha256_* APIs out of FreeRTOS and into platforms
The sha256 API was implemented in the FreeRTOS port layer. It used mbedtls
to calculate the hash, but different versions of mbedtls can be used with
FreeRTOS.

This PR moves the golioth_sys_sha256_* APIs out of FreeRTOS and into
ESP-IDF and ModusToolbox ports. This solves a function name issue with the
older version of mbedtls in ModusToolbox. It also makes it easier for
custom ports to implement their own hashing.

Signed-off-by: Mike Szczys <mike@golioth.io>