ci: pin github runners to ubuntu 24.04
The GitHub Actions cloud runner `ubuntu-latest` images are moving
from Ubuntu 22.04 LTS to Ubuntu 24.04 LTS. We should be pinning
these anyway, so we'll pin to Ubuntu 24.04 LTS. This also
suppresses a warning message in the GitHub UI.

The Modus Toolbox build workflow will remain pinned to Ubuntu 22.04
for now, as it does not build without modification on Ubuntu 24.04.
Ubuntu 22.04 is supported for another 2 years.

Signed-off-by: Sam Friedman <sam@golioth.io>