port: esp-idf: add ifndef for IN6_IS_ADDR_V4MAPPED(a)
With ESP-IDF v5.4, LWIP was upgraded to v2.2.0. This includes a new
IN6_IS_ADDR_V4MAPPED(a) macro that is defined in lwip/inet.h. It conflicts
with the macro in Golioth's copy of the ESP-IDF port's coap_config_posix.h.

The macro is still present in ESP-IDF's coap_config_posix.h, so instead
of removal, the macro in Golioth's coap_config_posix.h is guarded with
ifndef.

Signed-off-by: Mike Szczys <mike@golioth.io>