location: detect interchangable golioth_location_*_append() calls
golioth_location_*_append() is designed in a way that it cannot be called
interchangably for example with "cell" and "wifi" data. Instead one needs
to be finished before the other can be started.

Track which technologies (e.g. "cell" or "wifi") have been started and
finished. Do not allow to restart whatever has been finished.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>