ci: don't store artifacts while cancelling workflows
Prefer `success() || failure()` instead of `always()` for workflow
steps related to storing artifacts, and generating and storing test
reports. This ensures that those steps won't run when a workflow is
cancelled. This will allow runs to be cancelled faster, and also
prevents incomplete results from being included in test reports.

Signed-off-by: Sam Friedman <sam@golioth.io>