tests: settings: delay after connection for log flush
When registering settings during the HIL test, we get a flood
of log messages. On some platforms this can overwhelm the log
buffer and lead to dropped messages. If one of those dropped
messages is the connection message that the test scripts wait
for, the test will fail.

Now, we add a small delay after registering the settings to give
time for the log buffer to flush, and then send a new "settings
registration complete" message for the test script to key off of.
This ensures the test script receives the message, and
increases the stability of the test.

Signed-off-by: Sam Friedman <sam@golioth.io>