fw_update: refactor fw_update_validate to fw_update_check_candidate
Update fw_update_validate to fw_update_check_candidate and pass an image
hash and size such that image validateion is not coupled to any state
maintained in ports. Ports should be able to validate an image in a
candidate slot without needing to have downloaded since last reboot.

Make a corresponding update in fw_update thread to check whether the
requested firmware update image is already present in candidate slot. If
so, download is skipped and we attempt to reboot into existing candidate
slot image.

Note that the call to fw_update_validate is now removed from the post
download flow and not replaced in the same call site by
fw_update_check_image. fw_update_validate was previously a no-op for most
platform ports, and for those that did implement checks (esp_idf), the
functionality was moved to fw_update_post_download, which is still called
following an image download.

Signed-off-by: Daniel Mangum <georgedanielmangum@gmail.com>