tests: hil: scripts: package plugins in directories
We were previously packaging our HIL pytest plugins as collections
of python scripts without any directory hierarchy. This meant that
every python file in the package was placed in the root of the
site_packages directory. If any other package did the same and had
names that conflicted, the files would be overwritten. For example
these two plugins both contain a file named plugin.py and both
attempt to place that file at the root of the site packages
directory. Normal practice in python is to create a directory for
each package, and that's what's done in this commit.

Signed-off-by: Sam Friedman <sam@golioth.io>