hil: add NXP frdm_rw612 to integration tests
This board has a very large (64 MB) flash chip that takes between 4-8
minutes to fully erase. Because of this, the erase step in the HIL plugin
has been removed.

This is not an issue during sample tests as the erase is performed using
JLinkExe and the limited memory range reduces erase time to <30 seconds.
However, the pylink plugin we use for integration tests doesn't currently
support erasing a memory range.

Signed-off-by: Mike Szczys <mike@golioth.io>