zephyr: deprecate implicit use of 'golioth_esp_at' shield
nrf52840dk was always assumed to be used with ESP-AT WiFi companion chip.
This hard dependency was dropped with the introduction of 'golioth_esp_at'
shield. Deprecate implicit use of that shield (and the use of ESP-AT WiFi)
and prefer explicit use.

This will allow to use other shields or network technologies with existing
samples, on nrf52840dk board.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>