tests: hil: zephyr: move from boards/ to socs/ when appropriate
This is a followup to commit b7137d731aca ("examples: zephyr: move from
boards/ to socs/ when appropriate").

Some devicetree and configuration overlays are not board specific, but they
apply to SoC used for that board. Move such overlays to socs/ under
application directory, so they apply to wider range of boards
automatically.

Do this for following boards -> socs:
 * esp32_devkitc_wrover/esp32/procpu -> esp32/procpu
 * native_sim -> native
 * native_sim/native/64 -> native/64
 * nrf9160dk/nrf9160/ns -> nrf9160/ns

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>